Mention use of Mixins in README
I just wondered why all the methods were `undefined`. Thought that the addon would reopen `Ember.Component` and so on to add them automatically, but apparently that is not what is happening (until #1 lands). But having to apply the mixins by yourself was nowhere mentioned, also the examples were misleading in that regard. 

Added the missing pieces!

Super useful addon btw! :)